prep_fasta change and wrapper change
Changed the prep_fasta_and_memechip.sh to actually account for file directories in the wrapper script.

Also, added a ton of code for the wrapper script to first produce the enhancer/promoter bed files needed for meme-suite, which calls on prep_fasta_andmemechip.sh to produce the meme-suite files.